Aug 29 Commit 1 Restucture and move directories and paths
