+++ data_util.py
@@ -1,8 +1,11 @@
 # !/usr/local/bin/python
+from __future__ import print_function
 import numpy as np
+import librosa.display as lib_disp
+import librosa.feature as lib_feat
 
 def sparse_tuple_from(sequences, dtype=np.int32):
-    """Create a sparse representention of x.
+    """Create a sparse representention of x. For handling one-hot vector
     Args:
         sequences: a list of lists of type dtype where each element is a sequence
     Returns:
@@ -11,8 +14,8 @@ def sparse_tuple_from(sequences, dtype=np.int32):
     indices = []
     values = []
 
-    for n, seq in enumerate(sequences):
-        indices.extend(zip([n]*len(seq), range(len(seq))))
+    for i, seq in enumerate(sequences):
+        indices.extend(zip([i]*len(seq), range(len(seq))))
         values.extend(seq)
 
     indices = np.asarray(indices, dtype=np.int64)
@@ -20,6 +23,37 @@ def sparse_tuple_from(sequences, dtype=np.int32):
     shape = np.asarray([len(sequences), np.asarray(indices).max(0)[1]+1], dtype=np.int64)
 
     return indices, values, shape
+
+def features(rawsnd, num) :
+    """
+    Summary:
+        Compute audio features
+    Parameters:
+        rawsnd : array with string paths to .wav files
+        num : numbers of mfccs to compute
+    Output:
+        Return a (num+28,max_stepsize*32)-dimensional Tensorflow feature vector, and length in
+        *num Amount of Mel Frequency Ceptral Coefficients
+        *12 Chromagrams
+        *7 bands of Spectral Contrast
+        *6 bands of Tonal Centroid Features (Tonnetz)
+        *Zero Crossing Rate
+        *Spectral Rolloff
+        *Spectral Centroid
+    """
+    x, _ = librosa.load(rawsnd,sr=sample_rate, duration=max_stepsize)
+    s_tft = np.abs(librosa.stft(x))
+    ft = lib_feat.mfcc(y=x, sr=sample_rate, n_mfcc=num)
+    ft = np.append(ft,lib_feat.chroma_stft(S=s_tft, sr=sample_rate),axis=0)
+    ft = np.append(ft,lib_feat.spectral_contrast(S=s_tft,sr=sample_rate),axis=0)
+    ft = np.append(ft,lib_feat.tonnetz(y=librosa.effects.harmonic(x), sr=sample_rate),axis=0)
+    ft = np.append(ft,lib_feat.zero_crossing_rate(y=x),axis=0)
+    ft = np.append(ft,lib_feat.spectral_rolloff(y=x,sr=sample_rate),axis=0)
+    ft = np.append(ft,lib_feat.spectral_centroid(y=x,sr=sample_rate),axis=0)
+    z = np.zeros((num+12+7+6+3,(max_stepsize*((sample_rate/1000)*2))-ft.shape[1]))
+    ft = np.concatenate((ft,z),axis=1)
+    #print(ft[13].astype(np.float16))
+    return (ft)
 
 def load_dir(fp):
     """
@@ -44,7 +78,7 @@ def load_dir(fp):
                 if not ("SA" in __file):
                     ind+=1
                     if (ind%500==0):
-                        print(".", end="")
+                        print(".", end='')
                     if (".wav" in __file):
                         raw_audio.append(__file)
                     if (".PHN" in __file):
@@ -70,7 +104,7 @@ def load_dir(fp):
         return raw_audio,phonemes,words,text,ind
 
 def pad_seq():
-
+    return 0
 
 def getData(filepath, minibatch_size, batches ,num_features ,num_classes):
     """
