+++ tf_start
@@ -1,6 +1,3 @@
-#cd /home/jacky/Desktop/Spyre
-#source nsound/bin/activate
-#cd nsound_git
 python tf_model.py
 
 #sudo python -m tensorflow.tensorboard --logdir=/home/jacky/Desktop/Spyre/nsound_git/totalsummary --port 6006
